+++ tests/src/all_suites.adb
@@ -8,6 +8,7 @@ with Composition_Tests;
 with Collection_Tests;
 with Nested_Tests;
 with Deferred_Tests;
+with Diagnostics_Tests;
 
 package body All_Suites is
 
@@ -25,6 +26,7 @@ package body All_Suites is
       Collection_Tests.Add_Tests (Result);
       Nested_Tests.Add_Tests (Result);
       Deferred_Tests.Add_Tests (Result);
+      Diagnostics_Tests.Add_Tests (Result);
       return Result;
    end Suite;